Add source and project files for Hyper-V agent. As a general guid, *.cs files are source, *.csproj file is Visual Studio project that contains build settings and instructions, settings of all kinds are in XML, *.Designer.* files are generated by Visual Studio for its various editors.
More details...
packages.config references NuGet library packages that the project relies on.
Settings for the agent are held in AgentShell.settings in XML format
AgentShell.Designer.cs includes wrappers for easy access to the settings.
Service1 subytypes ServiceBase, which provides it with all the hooks to be run as a Windows Service, which is the Windows equivalent of a Daemon process.  The name is unfortunate, and may change.  The service uses a trick whereby it can be run as a Console application for debugging in addition to being capable of being run as a Windows Service.  The reason is that you cannot run a Windows Service direcdly.  It has to be started by the Service Control Manager (SCM), which gets in the way of developers.  The decision to run as a console application or a Windows Service is taken in Program.cs on the basis of a command line arguement.